+++ client/include/global.hpp
@@ -255,7 +255,8 @@ namespace HavocNamespace
             QString Port;
             QString User;
             QString Password;
-            bool    IgnoreSSLErrors = false;
+            bool    PasswordIsHashed = false;
+            bool    IgnoreSSLErrors  = false;
 
             std::vector<ListenerItem>      Listeners;
             std::vector<json>              RegisteredListeners;